changes to support time based bqm queries
